pyproject.toml
[tool.poetry]
name = "foliumroutesave"
version = "0.1.0"
description = ""
authors = ["cocomoff <keisuke.otaki.jp@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
jupyterlab = "^3.4.5"
numpy = "^1.23.2"
folium = "^0.12.1"
python-dotenv = "^0.20.0"
openrouteservice = "^2.3.3"
pyppeteer = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
